+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Added a lint group for Conditionals
+* Added a section for Conditionals
 * Added LinkedItem support
 * Added an unneeded LinkedItem lint
 * Added a lint to prevent nested items